requirements.txt
fastapi 
firebase-admin 
google-auth 
requests 
pydantic-settings 
python-dotenv 
uvicorn 
python-multipart 
email-validator